Move overclocks and frameworks data to data folder
