+++ src/pages/frameworks/FrameworksPage.tsx
@@ -1,9 +1,9 @@
 import { useCallback } from 'react';
+import { Frameworks } from 'data/frameworks';
 import MinerPageLayout from 'pages/MinerPageLayout';
 import useStore from 'store/useStore';
 import { Miner } from 'utils/miner';
 import { MinerWeapon, MinerWeapons } from 'utils/weapons';
-import { Frameworks } from './FrameworkData';
 import MinerFrameworks from './MinerFrameworks';
 
 export default function FrameworksPage() {
