+++ src/pages/overclocks/OverclockCard.tsx
@@ -1,10 +1,10 @@
 import InfoCircleOutlined from '@ant-design/icons/InfoCircleOutlined';
 import { Card, Col, Popover } from 'antd';
 import { isMobile } from 'react-device-detect';
+import { Overclock } from 'data/overclocks';
 import { Miner, MinerColor, MinerColorContrastText } from 'utils/miner';
 import { MinerWeapon } from 'utils/weapons';
 import OverclockCardPopover from './OverclockCardPopover';
-import { Overclock } from './OverclockData';
 import OverclockIcon from './OverclockIcon';
 
 export default function OverclockCard(props: {
