+++ src/store/state.ts
@@ -1,4 +1,4 @@
-import { Framework } from 'pages/frameworks/FrameworkData';
+import { Framework } from 'data/frameworks';
 import { Miner } from 'utils/miner';
 import { MinerWeapon } from 'utils/weapons';
